v5 code refactor + break-up + tidy
